requirements.txt
Flask==1.1.2
Pillow==9.4.0
requests==2.25.1
jinja2==2.11.3
MarkupSafe==1.1.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.0.1
future==0.18.2
glcontext==2.3.4
h11==0.14.0
httplib2==0.21.0
inflection==0.5.1
multidict==6.0.4
openai==0.26.5
rsa==4.9
text-unidecode==1.3
tokenizers==0.13.2
trio==0.22.0
uritemplate==4.1.1
webencodings==0.5.1
wsproto==1.2.0
yarl==1.8.2
itsdangerous==1.1.0
werkzeug==2.0.3
